net: Decouple DHCP server from converter pkg
The GetResolvConfDetailsFromPod function reads and parses
`/etc/resolv.conf` - the DNS resolver's configuration file.

Move it from `pkg/virt-launcher/virtwrap/converter` to
the `pkg/network/dns` package - under SIG network's ownership.
