+++ pkg/network/driver/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "//pkg/network/dhcp/server:go_default_library",
         "//pkg/network/dhcp/serverv6:go_default_library",
         "//pkg/network/dns:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/github.com/golang/mock/gomock:go_default_library",
