+++ pkg/virt-launcher/virtwrap/converter/BUILD.bazel
@@ -27,7 +27,6 @@ go_library(
         "//pkg/handler-launcher-com/cmd/v1:go_default_library",
         "//pkg/host-disk:go_default_library",
         "//pkg/ignition:go_default_library",
-        "//pkg/network/dns:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/pointer:go_default_library",
         "//pkg/storage/reservation:go_default_library",
